Vendor update
+ added vet and lint checks to travis script

+++ enc/hash.h
@@ -369,7 +369,7 @@ class HashLongestMatch {
       if (prev_ix >= cur_ix) {
         continue;
       }
-      if (PREDICT_FALSE(backward > max_backward)) {
+      if (PREDICT_FALSE(backward > (int)max_backward)) {
         continue;
       }
       prev_ix &= static_cast<uint32_t>(ring_buffer_mask);
@@ -510,7 +510,7 @@ class HashLongestMatch {
         if (len > kMaxZopfliLen) {
           matches = orig_matches;
         }
-        *matches++ = BackwardMatch(backward, len);
+        *matches++ = BackwardMatch(static_cast<int>(backward), len);
       }
     }
     const uint32_t key = HashBytes(&data[cur_ix_masked]);
